add file picker, file upload functionality and navigation to create request scree
Because:
- We want users to select a file from his device as a proof of him belonging to an organization
- We want users to upload that file to firebase-storage
- We only want users who have uploaded the proof document to create a request

Added buttons to select file and then for uploading file on firebase-storage.
Added a toast to display a waiting message. Percentage progress of upload task is shown below upload button.
Added a button which will be enabled only after a name is entered and file is fully uploaded and which onTap will navigate to create request screen.

For now only functionality is focus, UI will be improved later on.